split unit tests for persistence and models
